fix: align queue read count with pgmq integer type

+++ supabase/migrations/20260824000200_queues.sql
@@ -69,7 +69,7 @@ create or replace function public.queue_read(
 )
 returns table (
   id text,
-  read_count bigint,
+  read_count integer,
   enqueued_at timestamptz,
   visible_at timestamptz,
   message jsonb
